test : 지하철 역 생성 수정 삭제에 대한 테스트 추가

+++ src/main/java/wooteco/subway/station/application/StationService.java
@@ -3,6 +3,8 @@ package wooteco.subway.station.application;
 import org.springframework.dao.DuplicateKeyException;
 import org.springframework.stereotype.Service;
 import wooteco.subway.exception.badrequest.DuplicateNameException;
+import wooteco.subway.exception.badrequest.IllegalDeleteException;
+import wooteco.subway.line.dao.SectionDao;
 import wooteco.subway.station.dao.StationDao;
 import wooteco.subway.station.domain.Station;
 import wooteco.subway.station.dto.StationRequest;
@@ -13,10 +15,12 @@ import java.util.stream.Collectors;
 
 @Service
 public class StationService {
-    private StationDao stationDao;
+    private final StationDao stationDao;
+    private final SectionDao sectionDao;
 
-    public StationService(StationDao stationDao) {
+    public StationService(StationDao stationDao, SectionDao sectionDao) {
         this.stationDao = stationDao;
+        this.sectionDao = sectionDao;
     }
 
     public StationResponse saveStation(StationRequest stationRequest) {
@@ -24,7 +28,7 @@ public class StationService {
             Station station = stationDao.insert(stationRequest.toStation());
             return StationResponse.of(station);
         } catch (DuplicateKeyException e) {
-            throw new DuplicateNameException("이미 존재하는 역 이름입니다.");
+            throw new DuplicateNameException("이미 존재하는 지하철 역입니다");
         }
     }
 
@@ -41,13 +45,20 @@ public class StationService {
     }
 
     public StationResponse updateStation(Long id, StationRequest stationRequest) {
-        Station station = stationDao.findById(id);
-        Station updateStation = station.update(stationRequest.getName());
-        stationDao.update(updateStation);
-        return StationResponse.of(updateStation);
+        try {
+            Station station = stationDao.findById(id);
+            Station updateStation = station.update(stationRequest.getName());
+            stationDao.update(updateStation);
+            return StationResponse.of(updateStation);
+        } catch (DuplicateKeyException e) {
+            throw new DuplicateNameException("이미 존재하는 지하철 역입니다");
+        }
     }
 
     public void deleteStationById(Long id) {
+        if (sectionDao.existStation(id)) {
+            throw new IllegalDeleteException("이미 노선에 등록된 지하철 역입니다");
+        }
         stationDao.deleteById(id);
     }
 }
